requirements.txt
phidata
ollama
streamlit>=1.30.0
fastapi
uvicorn
pydantic>=2.0.0
openai>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
pandas>=2.0.0 